some modifications from instance

+++ executors/convert_executor_sqs.py
@@ -16,7 +16,7 @@ dbname = os.environ['RDS_DATABASE']
 folder = os.environ['PROCESS_FOLDER']
 bucket = os.environ['BUCKET']
 s3 = boto3.resource("s3")
-sqs = boto3.resource('sqs')
+sqs = boto3.client('sqs', region_name='us-east-1')
 engine = create_engine(f'postgresql://{user}:{password}@{hostname}/{dbname}')
 connection = engine.connect()
 session = Session(bind=connection)  # create a Session
@@ -39,24 +39,27 @@ class ConvertBySQS:
     if __name__ == "__main__":
         while True:
             messages = sqs.receive_message(QueueUrl=os.environ.get('QUEUE_URL'))
-            for msg in messages['Messages']:
-                msg_body = msg['Body']
-                task_id = int(msg_body)
-                t = get_task(task_id)
-                print(f'The message body: {msg_body}')
-                command = f'ffmpeg -i {folder}' + str(t.filename) + \
-                          f' {folder}' + t.filename[:-3] + str(t.newformat)
-                try:
-                    s3.download_file(bucket, t.filename, t.filename)
-                    subprocess.Popen(command, shell=True)
-                    s3.upload_fileobj(f'{folder}{t.filename[:-3]}{str(t.newformat)}', bucket,
-                                      f'{folder}{t.filename[:-3]}{str(t.newformat)}')
-                    if os.path.exists(folder + t.filename) and os.path.exists(
-                            f'{folder}{t.filename[:-3]}{str(t.newformat)}'):
-                        os.remove(folder + t.filename)
-                        os.remove(f'{folder}{t.filename[:-3]}{str(t.newformat)}')
-                    t.status = "PROCESSED"
-                    session.commit()
-                    print("Conversión realizada con exito")
-                except Exception as e:
-                    print(e)
+            print(messages)
+            if messages:
+              for msg in messages['Messages']:
+                  print(msg)
+                  msg_body = msg['Body']
+                  task_id = int(msg_body)
+                  t = self.get_task(task_id)
+                  print(f'The message body: {msg_body}')
+                  command = f'ffmpeg -i {folder}' + str(t.filename) + \
+                            f' {folder}' + t.filename[:-3] + str(t.newformat)
+                  try:
+                      s3.download_file(bucket, t.filename, t.filename)
+                      subprocess.Popen(command, shell=True)
+                      s3.upload_fileobj(f'{folder}{t.filename[:-3]}{str(t.newformat)}', bucket,
+                                        f'{folder}{t.filename[:-3]}{str(t.newformat)}')
+                      if os.path.exists(folder + t.filename) and os.path.exists(
+                              f'{folder}{t.filename[:-3]}{str(t.newformat)}'):
+                          os.remove(folder + t.filename)
+                          os.remove(f'{folder}{t.filename[:-3]}{str(t.newformat)}')
+                      t.status = "PROCESSED"
+                      session.commit()
+                      print("Conversión realizada con exito")
+                  except Exception as e:
+                      continue
